feat(database,grpc-sdk): requested changes
