+++ libraries/grpc-sdk/src/interfaces/index.ts
@@ -3,3 +3,4 @@ export * from './ConduitService';
 export * from './Model';
 export * from './GrpcCallback';
 export * from './Indexable';
+export * from './RawQuery';
