+++ libraries/grpc-sdk/src/types/db.ts
@@ -31,28 +31,3 @@ export type Query<T> =
   | setQuery<T>
   // | arrayQuery<T>
   | conditionalQuery<T>;
-
-export interface RawQuery {
-  mongoQuery?: RawMongoQuery;
-  sqlQuery?: RawSQLQuery;
-}
-
-export type RawMongoQuery = {
-  aggregate?: object[];
-  count?: object;
-  deleteOne?: object;
-  deleteMany?: object | object[];
-  drop?: object;
-  explain?: object;
-  find?: object;
-  insertOne?: object;
-  insertMany?: object | object[];
-  updateOne?: object;
-  updateMany?: object | object[];
-  options?: object;
-};
-
-export type RawSQLQuery = {
-  query: string;
-  options?: object;
-};
